Removed the TourSteps.js(Reason : Duplicate code)
Removed the TourSteps.js file and merged the new arrow color changes code in TourSteps.jsx file
